double to float and fixes

+++ src/Slim/planar/PlanarSpeedModule.hpp
@@ -3,7 +3,7 @@
 
 #include "../OrientationSensorsWrapper.hpp"
 
-extern Point3D<double> speed_vector;
+extern Point3D<float> speed_vector;
 
 void initPlanarSpeed(OrientationSensors&);
 void updatePlanarSpeed(OrientationSensors&);
